omega: widen literal-fold bit-mask check past the last u16 rule bit
91b7bba9 added SATURATING_ADD_ZERO_BIT at 1<<15, filling all sixteen
KNOWN_BITS and making the unknown-bit mask a vacuous u16 test — clippy
bad_bit_mask turned the crate red. Widen the check to u32 so it still
fires if a bit is ever retired.

+++ omega-rust/omega/pipeline/selected-instructions-to-selected-instructions/src/rewrites/selected_lowering/literal_fold/model.rs
@@ -289,7 +289,7 @@ impl LiteralFoldPolicy {
     }
 
     pub const fn from_canonical_bits(bits: u16) -> Option<Self> {
-        if bits == 0 || bits & !Self::KNOWN_BITS != 0 {
+        if bits == 0 || (bits as u32) & !(Self::KNOWN_BITS as u32) != 0 {
             None
         } else {
             Some(Self {
